backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3

# LLM providers
anthropic==0.40.0
openai==1.57.4
google-generativeai==0.8.3

# Data skills
requests==2.32.3
yfinance==1.2.0
finvizfinance==0.14.6

# Scraping
playwright>=1.44.0
beautifulsoup4==4.12.3
lxml==6.0.2
curl_cffi>=0.7.0

# Utilities
httpx==0.28.1
